+++ ios/Classes/PhoneEntryViewFactory.swift
@@ -14,7 +14,12 @@ final class PhoneEntryViewFactory: NSObject, FlutterPlatformViewFactory {
         viewIdentifier viewId: Int64,
         arguments args: Any?
     ) -> FlutterPlatformView {
-        PhoneEntryPlatformView(frame: frame, viewIdentifier: viewId, messenger: messenger)
+        PhoneEntryPlatformView(
+            frame: frame,
+            viewIdentifier: viewId,
+            messenger: messenger,
+            args: args
+        )
     }
 
     func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {